refactor: tweaking controllers and folder structure
