+++ app/controllers/disciplineControllers.js
@@ -1,9 +1,9 @@
-const database = require('../models/database');
+const database = require('../../config/database');
 
 exports.registerDisciplines = (req, res) => {
-    const { nome, fase, cargaHoraria } = req.body;
-    const query = 'INSERT INTO disciplinas (nome_materia, fase_materia, carga_horaria) VALUES($1, $2, $3);'
-    const values = [nome, fase, cargaHoraria]
+    const { nome, idFase } = req.body;
+    const query = 'INSERT INTO disciplinas (nome_materia, id_fase ) VALUES($1, $2);'
+    const values = [nome, idFase]
     database.query(query, values).then(
         () => {
             res.status(201).send({ mensagem: 'Discipline successfully registered.' });
@@ -43,13 +43,12 @@ exports.deleteDisciplines = (req, res) => {
 }
 
 exports.updateDisciplines = (req, res) => {
-    const {id} = req.params;
-    const { nome, fase, cargaHoraria } = req.body;
-    console.log(nome, fase, cargaHoraria, id)
-    const query = 'UPDATE disciplinas SET nome_materia=$1, fase_materia=$2, carga_horaria =$3 WHERE id_materia=$4;';
-    const values = [nome, fase, cargaHoraria, id];
+    const { id } = req.params;
+    const { nome, idFase } = req.body;
+    const query = 'UPDATE disciplinas SET nome_materia=$1, id_fase=$2 WHERE id_materia=$3;';
+    const values = [nome, idFase, id];
     database.query(query, values).then(
-        ()=> {
+        () => {
             res.status(200).send({ mensagem: 'Discipline successfully updated.' });
         },
         (erro) => {
@@ -57,3 +56,6 @@ exports.updateDisciplines = (req, res) => {
         }
     )
 }
+
+
+
