+++ app/routes/profRoutes.js
@@ -6,6 +6,6 @@ router
 .post('/', controller.registerProfessor)
 .delete('/:id', controller.deleteProfessor)
 .get('/', controller.listProfessor)
-.put('/:nome', controller.updateProfessor);
+.put('/:id', controller.updateProfessor);
 
 module.exports = router;
