+++ server.js
@@ -4,13 +4,13 @@ const PORT = 3000;
 const app = express();
 app.use(bodyParser.json());
 
-const professorRoutes = require('./routes/profRoutes');
+const professorRoutes = require('./app/routes/profRoutes');
 app.use('/professor', professorRoutes);
 
-const disciplineRoutes = require('./routes/disciplineRoutes');
+const disciplineRoutes = require('./app/routes/disciplineRoutes');
 app.use('/disciplina', disciplineRoutes);
 
-const classroomsRoutes = require('./routes/classroomRoutes');
+const classroomsRoutes = require('./app/routes/classroomRoutes');
 app.use('/sala', classroomsRoutes);
 
 app.listen(PORT, () => {
